// orbe-ledger/src/infra/database/common/enums/idempotency.status.ts
export enum IdempotencyStatus {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  FAILED = 'FAILED',
  COMPLETED = 'COMPLETED',
}

export enum EntityType {
  ACCOUNT_TYPE = 'ACCOUNT_TYPE',
  ACCOUNT = 'ACCOUNT',
  AUDIT = 'AUDIT',
  BALANCE_SNAPSHOT = 'BALANCE_SNAPSHOT',
  CHART_OF_ACCOUNTS = 'CHART_OF_ACCOUNTS',
  CURRENCY = 'CURRENCY',
  ENTRY = 'ENTRY',
  HOLD = 'HOLD',
  IDEMPOTENCY = 'IDEMPOTENCY',
  INBOX = 'INBOX',
  JOURNAL = 'JOURNAL',
  LEDGER = 'LEDGER',
  LIMIT = 'LIMIT',
  ORGANIZATION = 'ORGANIZATION',
  OUTBOX = 'OUTBOX',
  RECONCILIATION = 'RECONCILIATION',
  SAGA = 'SAGA',
  TRANSACTION = 'TRANSACTION',
}